# Fix memory leak in `Craft._groupParamsByDeltaNames`. 