### Description

See #9575


### Benchmark
*Before*
```
644.090087890625 ms
650.367919921875 ms
659.84326171875 ms
660.115966796875 ms
635.376953125 ms
655.094970703125 ms
651.10205078125 ms
653.590087890625 ms
651.787109375 ms
652.337158203125 ms
645.884033203125 ms
652.279052734375 ms
642.240966796875 ms
649.8369140625 ms
654.277099609375 ms
```

*After*
```
26.755859375 ms
37.947998046875 ms
41.156005859375 ms
38.9599609375 ms
19.1240234375 ms
42.887939453125 ms
20.4990234375 ms
22.829833984375 ms
21.895751953125 ms
21.466064453125 ms
24.864990234375 ms
21.7001953125 ms
21.502197265625 ms
19.807861328125 ms
20.426025390625 ms
19.064208984375 ms
```