@brandonkelly 2021-07-20
Thanks! Just released 3.7.4 with this change.